% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Draw a name}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{is a name}

\item{the_sex}{is either "F" or "M"}
}
\value{
the graph of the popularity of the name accross history
}
\description{
Draw a name
}
