% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Draw names}
\usage{
draw_names(names)
}
\arguments{
\item{names}{a vector of names}
}
\value{
the graph of the popularity of your names
}
\description{
Draw names
}
