% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{data}
\name{project}
\alias{project}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 1000 rows and 5 variables:
\describe{
\item{ID}{personal identitfication number}
\item{gender}{gender of individual}
\item{height}{height, in cm}
\item{weight}{weight, in kg}
\item{phys}{amount of physical activity}
...
}
}
\usage{
project
}
\description{
A datas et containing the ID, gender, height, weight, and amount of physical activity.
}
\keyword{datasets}
