% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save as csv}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{is a dataset}

\item{filename}{is a file name}

\item{row.names}{is FALSE}

\item{...}{is what you want}
}
\value{
a csv file
}
\description{
Save as csv
}
